streamlit/requirements.txt
streamlit==1.9.0
requests==2.27.1
Pillow==9.4.0